requirements.txt
langchain==0.0.101
streamlit==0.87.0
openai==0.13.0
python-dotenv==0.19.2
pinecone-client==2.0.0
scikit-learn==1.0.2
sentence-transformers==2.1.0
